feat: create Machine model for backend with properties and relationships
This commit introduces a new Machine model in the backend, defining properties such as Id, Name, SerialNumber, Type, InstallationDate, IsActive, and Description. It establishes a relationship with ProductionData, enhancing the data structure for managing machine-related information.
