+++ mobile/Models/Machine.cs
@@ -6,7 +6,10 @@ public class Machine
 {
     public int Id { get; set; }
     public string Name { get; set; }
+    public string Manufacturer { get; set; }
+    public string Model { get; set; }
     public string SerialNumber { get; set; }
-    public string Type { get; set; }
     public bool IsActive { get; set; }
+    public DateTime InstallationDate { get; set; }
+    public string? Description { get; set; }
 }
